02_Prompt_Engineering/requirements.txt
openai
anthropic
google-generativeai
python-dotenv 